Latest (#604)
* * fix up some Linux compilation warnings - subtext is array and always non-null, missing default case, snprintf missing "%s" arg, Text.cpp missing extra %c in textAndFont parsing

* * dropdown widget for hud

* * action prompts unique skill lvl backing
* dropdown has invisible click blocking frame behind it
* animate cursor for dropdown
* /dropgold uses amount parameter - bounds checking in net.cpp
* close GUI dropdowns on status open/close
* skillsheet not interactable when moving

* * bounds checking on dropgold command
* add map messages to character sheet

* * stat tooltips for character sheet
* getBaseManaRegen no longer part of Entity
* make more frames hollow to avoid blocking clicks

* * charisma when trading now uses the bonus value from equipment

* * Item tooltips use Pixel Maz
* Add method to color highlight words if text has touching outlines in Field() + Text()
* Most item tooltips use setTextColor, not setColor for multicolor text

* * remove unused constructor for Text when doing highlighting

* * new right click menus, add a block click frame while active
* right click menu adjust left/right based on compact layout
* right click menu adjust to not fall off top/bottom edges of screen and offset mouse position
* backpack slots fix compact view position

* * lang update

* * larger pixelmix header for item tooltips
* undo debug lines on tooltip if not debugging

* * debug timer class to add time logging anywhere

* * add right click for dropdown inputs

* * reduce findFrame calls in skillsheet for a little bit of optimisation

+++ src/draw.cpp
@@ -2548,7 +2548,7 @@ SDL_Rect ttfPrintTextColor( TTF_Font* font, int x, int y, Uint32 color, bool out
 		SDL_FreeSurface(textSurf);
 		// load the text outline surface as a GL texture
 		allsurfaces[imgref] = surf;
-		allsurfaces[imgref]->userdata = (void*)imgref;
+		allsurfaces[imgref]->userdata = (void*)((long int)imgref);
 		glLoadTexture(allsurfaces[imgref], imgref);
 		imgref++;
 		// store the surface in the text surface cache
